+++ apps/frontend/src/libs/components/header/header.tsx
@@ -5,8 +5,8 @@ import { AppRoute } from "~/libs/enums/enums.js";
 import { type ValueOf } from "~/libs/types/types.js";
 import { type UserAuthResponseDto } from "~/modules/users/libs/types/types.js";
 
-import { Button, TextLink } from "../components.js";
-import { AuthenticatedHeader } from "./libs/components/authenticated-header/authenticated-header.js";
+import { Button, Link } from "../components.js";
+import { AuthenticatedHeader } from "./libs/components/components.js";
 import styles from "./styles.module.css";
 
 type Properties = {
@@ -35,7 +35,7 @@ const Header = ({ actions, user }: Properties): JSX.Element => {
 	return (
 		<header className={styles["header"]}>
 			<div className={styles["container"]}>
-				<TextLink to={AppRoute.ROOT}>
+				<Link to={AppRoute.ROOT}>
 					<span className="visually-hidden">Go to landing</span>
 					<img
 						alt="SmartScapes Logo"
@@ -44,7 +44,7 @@ const Header = ({ actions, user }: Properties): JSX.Element => {
 						src={appLogo}
 						width={136}
 					/>
-				</TextLink>
+				</Link>
 				{renderHeaderContent()}
 			</div>
 		</header>