+++ apps/frontend/src/libs/components/icon/libs/maps/icon-name-to-svg.map.ts
@@ -8,10 +8,12 @@ import Edit from "~/assets/images/icons/edit.svg?react";
 import Error from "~/assets/images/icons/error.svg?react";
 import EyeOff from "~/assets/images/icons/eye-off.svg?react";
 import Eye from "~/assets/images/icons/eye.svg?react";
+import Link from "~/assets/images/icons/link.svg?react";
 import Logout from "~/assets/images/icons/logout.svg?react";
 import Map from "~/assets/images/icons/map.svg?react";
 import Message from "~/assets/images/icons/message.svg?react";
 import Route from "~/assets/images/icons/route.svg?react";
+import Search from "~/assets/images/icons/search.svg?react";
 import Tag from "~/assets/images/icons/tag.svg?react";
 import Trash from "~/assets/images/icons/trash.svg?react";
 import User from "~/assets/images/icons/user.svg?react";
@@ -31,10 +33,12 @@ const iconNameToSvg: Record<
 	error: Error,
 	eye: Eye,
 	eyeOff: EyeOff,
+	link: Link,
 	logout: Logout,
 	map: Map,
 	message: Message,
 	route: Route,
+	search: Search,
 	tag: Tag,
 	trash: Trash,
 	user: User,