+++ apps/frontend/src/libs/components/textarea/libs/constants/default-rows-count.constant.ts
@@ -1,0 +1,3 @@
+const DEFAULT_ROWS_COUNT = 7;
+
+export { DEFAULT_ROWS_COUNT };