+++ packages/shared/src/index.ts
@@ -71,6 +71,7 @@ export {
 	type PermissionItemDto,
 	PermissionKey,
 } from "./modules/permission/permission.js";
+export { type PlannedPathResponseDto } from "./modules/planned-paths/planned-paths.js";
 export {
 	type PointOfInterestPatchRequestDto,
 	pointOfInterestUpdateValidationSchema,
@@ -85,8 +86,11 @@ export {
 export {
 	type ReviewCreatePayload,
 	reviewCreateValidationSchema,
+	type ReviewGetAllSearchQuery,
+	reviewGetAllSearchQueryValidationSchema,
 	type ReviewGetByIdResponseDto,
 	type ReviewRequestDto,
+	ReviewsApiPath,
 } from "./modules/reviews/reviews.js";
 export {
 	type RouteConstructRequestDto,
@@ -110,12 +114,26 @@ export {
 	type UserUnfollowsParametersDto,
 } from "./modules/user-follows/user-follows.js";
 export {
+	UserRouteApiPath,
+	type UserRouteCreateRequestDto,
+	userRouteCreateValidationSchema,
+	type UserRouteParameters,
+	userRouteParametersValidationSchema,
+	type UserRoutePatchRequestDto,
+	userRoutePatchValidationSchema,
+	type UserRouteResponseDto,
+	UserRouteStatus,
+	type UserRouteStatusType,
+} from "./modules/user-routes/user-routes.js";
+export {
 	type AuthenticatedUserPatchRequestDto,
 	type AuthenticatedUserPatchResponseDto,
 	authenticatedUserPatchValidationSchema,
 	type UserAuthResponseDto,
 	UserExceptionMessage,
 	type UserGetByIdItemResponseDto,
+	userGetProfileParametersValidationSchema,
+	type UserPublicProfileResponseDto,
 	UsersApiPath,
 	type UserSignInRequestDto,
 	type UserSignInResponseDto,